requirements.txt
github3.py
cffi==1.11.5
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.17
pygit2==0.27.3
PyYAML==3.13
six==1.12.0
